backend/pyproject.toml
[tool.poetry]
name = "core"
version = "0.1.0"
description = ""
authors = ["Jean-Baptiste Truffault <jbtruffault.pro@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
asgiref = "3.8.1"
attrs = "23.1.0"
black = "23.10.0"
boto3 = "1.28.63"
botocore = "1.31.63"
celery = "5.4.0"
certifi = "2023.7.22"
cffi = ">=1.17.1"
charset-normalizer = "3.3.0"
click = "8.1.7"
colorama = { version = "0.4.6", markers = "sys_platform == 'win32' or platform_system == 'Windows'" }
cryptography = "41.0.4"
defusedxml = "0.7.1"
dj-database-url = "2.1.0"
django = "5.1.2"
django-cors-headers = "4.3.0"
django-filter = "24.3"
django-ses = "3.5.0"
django-storages = "1.14.2"
django-templated-mail = "1.1.1"
djangorestframework = "3.14.0"
djangorestframework-simplejwt = "5.3.1"
djoser = "2.2.0"
drf-spectacular = "0.26.5"
factory-boy = "3.3.0"
faker = "20.1.0"
flower = "2.0.1"
gevent = "24.11.1"
gpxpy = "1.6.2"
gunicorn = "21.2.0"
idna = "3.4"
inflection = "0.5.1"
iniconfig = "2.0.0"
isort = "5.12.0"
jmespath = "1.0.1"
joblib = "1.3.2"
jsonfield = "3.1.0"
jsonschema = "4.19.1"
jsonschema-specifications = "2023.7.1"
kombu = "5.4.2"
mypy-extensions = "1.0.0"
nltk = "3.8.1"
oauthlib = "3.2.2"
packaging = "23.2"
pastel = "0.2.1"
pathspec = "0.11.2"
pillow = "11.0.0"
platformdirs = "3.11.0"
pluggy = "1.3.0"
poethepoet = "0.24.1"
psycopg2-binary = "2.9.9"
pycparser = "2.21"
pyjwt = "2.8.0"
python-dateutil = "2.8.2"
python-decouple = "3.8"
python-dotenv = "1.0.0"
python3-openid = "3.2.0"
pytz = "2023.3.post1"
pyyaml = "6.0.1"
redis = "5.2.0"
referencing = "0.30.2"
regex = "2023.10.3"
requests = "2.31.0"
requests-oauthlib = "1.3.1"
rpds-py = "0.10.6"
ruff = "0.1.1"
s3transfer = "0.7.0"
six = "1.16.0"
social-auth-app-django = "5.3.0"
social-auth-core = "4.4.2"
sqlparse = "0.4.4"
tomli = "2.0.1"
tqdm = "4.66.1"
typing-extensions = "4.8.0"
uritemplate = "4.1.1"
urllib3 = "2.0.6"


[build-system]
requires = ["poetry-core>=1.5.0"]
build-backend = "poetry.core.masonry.api"

[tool.poe.tasks]
run = "poetry run python manage.py runserver"
test = "poetry run pytest ."
format = "poetry run black . "
lint = "poetry run ruff . --fix"
sort = "poetry run isort . "
precommit = ["format", "lint", "test"]
manage = "poetry run python manage.py"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.settings"
# -- recommended but optional:
python_files = ["test_*.py", "*_test.py", "testing/python/*.py"]